refactor(async): consolidate async client implementation
