--- src/ksef2/_async/__init__.py
@@ -1,1 +1,0 @@
-"""Async implementations — the single source of truth for the SDK."""
