--- src/ksef2/_async/core/__init__.py
@@ -1,9 +1,0 @@
-"""Async core infrastructure: transport, protocols, middlewares, mixins."""
-
-from ksef2._async.core.protocols import AsyncMiddleware
-from ksef2._async.core.http import AsyncHttpTransport
-
-__all__ = [
-    "AsyncHttpTransport",
-    "AsyncMiddleware",
-]
