--- src/ksef2/_async/core/types.py
@@ -1,17 +1,0 @@
-"""Shared type aliases for async transport, middleware, and endpoint layers."""
-
-from ksef2.core.types import (
-    Headers,
-    HeadersInput,
-    JsonObject,
-    QueryParamValue,
-    QueryParamsInput,
-)
-
-__all__ = [
-    "Headers",
-    "HeadersInput",
-    "JsonObject",
-    "QueryParamValue",
-    "QueryParamsInput",
-]
