--- src/ksef2/_async/endpoints/__init__.py
@@ -1,1 +1,0 @@
-"""Async endpoints package."""
