+++ src/ksef2/core/middlewares/async_auth.py
@@ -1,10 +1,10 @@
-from collections.abc import Mapping
-from typing import Any, final
+from typing import final, override
 
 import httpx
 
-from ksef2.core.middlewares.async_base import AsyncBaseMiddleware
 from ksef2.core.async_protocols import AsyncMiddleware
+from ksef2.core.middlewares.async_base import AsyncBaseMiddleware
+from ksef2.core.types import Headers, JsonObject, QueryParamsInput
 
 
 @final
@@ -13,20 +13,21 @@ class AsyncBearerTokenMiddleware(AsyncBaseMiddleware):
         self._next = transport
         self._token = token
 
-    def _merge(self, extra: dict[str, str] | None) -> dict[str, str]:
+    def _merge(self, extra: Headers | None) -> Headers:
         headers = {"Authorization": f"Bearer {self._token}"}
         return headers | (extra or {})
 
+    @override
     async def request(
         self,
         method: str,
         path: str,
         *,
-        headers: dict[str, str] | None = None,
-        params: Mapping[str, Any] | None = None,
-        json: dict[str, Any] | None = None,
+        headers: Headers | None = None,
+        params: QueryParamsInput | None = None,
+        json: JsonObject | None = None,
         content: bytes | None = None,
-        **kwargs,
+        **kwargs: object,
     ) -> httpx.Response:
         return await self._next.request(
             method,
