+++ src/ksef2/core/middlewares/async_base.py
@@ -1,8 +1,8 @@
 import abc
-from collections.abc import Mapping
-from typing import Any
 
 import httpx
+
+from ksef2.core.types import Headers, JsonObject, QueryParamsInput
 
 
 class AsyncBaseMiddleware(abc.ABC):
@@ -12,9 +12,9 @@ class AsyncBaseMiddleware(abc.ABC):
         method: str,
         path: str,
         *,
-        headers: dict[str, str] | None = None,
-        params: Mapping[str, Any] | None = None,
-        json: dict[str, Any] | None = None,
+        headers: Headers | None = None,
+        params: QueryParamsInput | None = None,
+        json: JsonObject | None = None,
         content: bytes | None = None,
     ) -> httpx.Response: ...
 
@@ -22,8 +22,8 @@ class AsyncBaseMiddleware(abc.ABC):
         self,
         path: str,
         *,
-        headers: dict[str, str] | None = None,
-        params: Mapping[str, Any] | None = None,
+        headers: Headers | None = None,
+        params: QueryParamsInput | None = None,
     ) -> httpx.Response:
         return await self.request("GET", path, headers=headers, params=params)
 
@@ -31,9 +31,9 @@ class AsyncBaseMiddleware(abc.ABC):
         self,
         path: str,
         *,
-        headers: dict[str, str] | None = None,
-        params: Mapping[str, Any] | None = None,
-        json: dict[str, Any] | None = None,
+        headers: Headers | None = None,
+        params: QueryParamsInput | None = None,
+        json: JsonObject | None = None,
         content: bytes | None = None,
     ) -> httpx.Response:
         return await self.request(
@@ -49,7 +49,7 @@ class AsyncBaseMiddleware(abc.ABC):
         self,
         path: str,
         *,
-        headers: dict[str, str] | None = None,
-        params: Mapping[str, Any] | None = None,
+        headers: Headers | None = None,
+        params: QueryParamsInput | None = None,
     ) -> httpx.Response:
         return await self.request("DELETE", path, headers=headers, params=params)
